+++ plugin/src/integrationTest/kotlin/net/bnb1/kradle/features/KtlintBlueprintTests.kt
@@ -1,7 +1,11 @@
 package net.bnb1.kradle.features
 
+import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import net.bnb1.kradle.PluginSpec
+import org.gradle.testkit.runner.TaskOutcome
+import org.gradle.testkit.runner.UnexpectedBuildFailure
 
 class KtlintBlueprintTests : PluginSpec({
 
@@ -34,5 +38,53 @@ class KtlintBlueprintTests : PluginSpec({
         val result = runTask("dependencies", "--configuration", "ktlint")
 
         result.output shouldContain "com.pinterest:ktlint"
+    }
+
+    test("Fail with no-semi rule") {
+        bootstrapProject {
+            """
+            jvm {
+                kotlin {
+                    lint {
+                        ktlint {
+                            rules {
+                                enable("no-semi")
+                            }
+                        }
+                    }
+                }
+                lint.enable()
+            }
+            """.trimIndent()
+        }
+        writeAppKt("println(\"Hello World!\");")
+
+        val ex = shouldThrow<UnexpectedBuildFailure> { runTask("ktlintMainSourceSetCheck") }
+
+        ex.message shouldContain "Unnecessary semicolon"
+    }
+
+    test("Succeed without error-prone rule set") {
+        bootstrapProject {
+            """
+            jvm {
+                kotlin {
+                    lint {
+                        ktlint {
+                            rules {
+                                disable("no-semi")
+                            }
+                        }
+                    }
+                }
+                lint.enable()
+            }
+            """.trimIndent()
+        }
+        writeAppKt("println(\"Hello World!\");")
+
+        val result = runTask("ktlintMainSourceSetCheck")
+
+        result.task(":ktlintMainSourceSetCheck")!!.outcome shouldBe TaskOutcome.SUCCESS
     }
 })