+++ plugin/src/integrationTest/kotlin/net/bnb1/kradle/features/PmdBlueprintTests.kt
@@ -1,10 +1,12 @@
 package net.bnb1.kradle.features
 
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.file.shouldExist
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import net.bnb1.kradle.PluginSpec
 import org.gradle.testkit.runner.TaskOutcome
+import org.gradle.testkit.runner.UnexpectedBuildFailure
 
 class PmdBlueprintTests : PluginSpec({
 
@@ -53,5 +55,53 @@ class PmdBlueprintTests : PluginSpec({
         val result = runTask("dependencies", "--configuration", "kradlePmd")
 
         result.output shouldContain "net.sourceforge.pmd:pmd-java"
+    }
+
+    test("Fail with error-prone rule set") {
+        bootstrapProject {
+            """
+            jvm {
+                java {
+                    codeAnalysis {
+                        pmd {
+                            ruleSets {
+                                errorProne(true)
+                            }
+                        }
+                    }
+                }
+                codeAnalysis.enable()
+            }
+            """.trimIndent()
+        }
+        writeAppJava("if (true) return;")
+
+        val ex = shouldThrow<UnexpectedBuildFailure> { runTask("pmdMain") }
+
+        ex.message shouldContain "UnconditionalIfStatement"
+    }
+
+    test("Succeed without error-prone rule set") {
+        bootstrapProject {
+            """
+            jvm {
+                java {
+                    codeAnalysis {
+                        pmd {
+                            ruleSets {
+                                errorProne(false)
+                            }
+                        }
+                    }
+                }
+                codeAnalysis.enable()
+            }
+            """.trimIndent()
+        }
+        writeAppJava("if (true) return;")
+
+        val result = runTask("pmdMain")
+
+        result.task(":pmdMain")!!.outcome shouldBe TaskOutcome.SUCCESS
     }
 })