+++ plugin/src/main/kotlin/net/bnb1/kradle/Configurable.kt
@@ -1,0 +1,9 @@
+package net.bnb1.kradle
+
+interface Configurable<T : Configurable<T>> {
+
+    operator fun invoke(action: T.() -> Unit = {}) = configure(action)
+
+    @Suppress("UNCHECKED_CAST")
+    fun configure(action: T.() -> Unit = {}) = action(this as T)
+}