+++ plugin/src/main/kotlin/net/bnb1/kradle/presets/KotlinJvmApplicationPreset.kt
@@ -17,6 +17,13 @@ class KotlinJvmApplicationPreset(project: Project) : Preset(project) {
             jvm.configureOnly {
                 kotlin {
                     useCoroutines()
+                    lint {
+                        ktlint {
+                            rules {
+                                disable("no-wildcard-imports")
+                            }
+                        }
+                    }
                     test {
                         useKotest()
                         useMockk()